% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_factor_labels.R
\name{apply_factor_labels}
\alias{apply_factor_labels}
\title{apply_factor_labels}
\usage{
apply_factor_labels(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
