% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age_at_visit.R
\name{calculate_age_at_visit}
\alias{calculate_age_at_visit}
\title{calculate_age_at_visit}
\usage{
calculate_age_at_visit(data, id, timepoint, birthdate, visitdate)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION}

\item{timepoint}{PARAM_DESCRIPTION}

\item{birthdate}{PARAM_DESCRIPTION}

\item{visitdate}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
