% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_predicted_probabilities.R
\name{calculate_predicted_probabilities}
\alias{calculate_predicted_probabilities}
\title{calculate_predicted_probabilities}
\usage{
calculate_predicted_probabilities(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{join_by}}
}
