% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_daily_equivalents.R
\name{convert_to_daily_equivalents}
\alias{convert_to_daily_equivalents}
\title{convert_to_daily_equivalents}
\usage{
convert_to_daily_equivalents(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
