% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diet_frequency_table}
\alias{diet_frequency_table}
\title{diet_frequency_table}
\usage{
diet_frequency_table()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\keyword{internal}
