% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cereal_daily_equivalents.R
\name{get_cereal_daily_equivalents}
\alias{get_cereal_daily_equivalents}
\title{FUNCTION_TITLE}
\usage{
get_cereal_daily_equivalents(
  data,
  cereal_names = c("cereal_type1", "cereal_type2")
)
}
\arguments{
\item{data}{get_cereal_daily_equivalents}

\item{cereal_names}{PARAM_DESCRIPTION, Default: c("cereal_type1", "cereal_type2")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{join_by}}
\code{\link[rlang]{sym}}
}
