% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncidsq-package.R
\docType{package}
\name{ncidsq-package}
\alias{ncidsq}
\alias{ncidsq-package}
\title{ncidsq: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
