% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion_size_adjust.R
\name{portion_size_adjust}
\alias{portion_size_adjust}
\title{portion_size_adjust}
\usage{
portion_size_adjust(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{join_by}}
}
