% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_constants}
\alias{remove_constants}
\title{remove_constants}
\usage{
remove_constants(environment = .GlobalEnv)
}
\arguments{
\item{environment}{PARAM_DESCRIPTION, Default: .GlobalEnv}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
