% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_variables}
\alias{rename_variables}
\title{rename_variables}
\usage{
rename_variables(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
