% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieve_norms}
\alias{retrieve_norms}
\title{retrieve_norms}
\usage{
retrieve_norms(
  gender = c("male", "female"),
  type = c("nhanes", "recommendations")
)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\keyword{internal}
