% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_r_data_from_excel}
\alias{update_r_data_from_excel}
\title{update_r_data_from_excel}
\usage{
update_r_data_from_excel(xlsx_directory)
}
\arguments{
\item{xlsx_directory}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[readxl]{read_excel}}
}
\keyword{internal}
